pyproject.toml
[project]
name = "zoo_concierge"
version = "1.0.0"
description = "Orchestrates personalized zoo experiences."
authors = [
    { name = "Linda Nichols", email = "lynnaloo@gmail.com" }
]
license = "Apache License 2.0"
readme = "README.md"
requires-python = ">=3.11"
package-mode = false

[tool.poetry.dependencies]
google-cloud-aiplatform = { extras = [
        "adk",
        "agent-engines",
], version = "^1.108.0" }
llama-index = "^0.12"
python = "^3.11"
pydantic = "^2.10.6"
pydantic-settings = "^2.8.1"
python-dotenv = "^1.0.1"
google-genai = "^1.16.1"
google-adk = "^1.0.0"
tabulate = "^0.9.0"
google-auth = "^2.36.0"
requests = "^2.32.3"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
scikit-learn = "^1.6.1"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"
cloudpickle = "^3.1.1"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"